Working and tested multicore version with time slices for a single channel
git-svn-id: svn+ssh://jop91.astron.nl/auto/home/jive_cvs/SOFTC/svn/trunk/sfxc@285 5736e3ea-d426-0410-8aaf-93e336ac3372

+++ include/genFunctions.h
@@ -24,5 +24,5 @@ int str2int(char *val, int& sval);
 
 long str_to_long (std::string inString, int pos, int length);
 
-void set_seed(uint32_t seed_);
+void set_seed(unsigned long seed_);
 int irbit2();
